refactor(web): refactor assets hooks
